Hourly FAO-56 and ASCE-EWRI Reference Evapotranspiration

The function calculates hourly or less than hour FAO-56 (grass) and ASCE-EWRI Reference (alfalfa) Evapotranspiration (ETo ETr). The "surface" parameter determines #' the type of ET. The function can also be used to scale hourly or less than hour actual evapotranspiration. The daily ET is implemented in ETo.

Usage

ETohr(
  Tmean,
  RH,
  DOY,
  Lz,
  t1,
  time,
  Lm,
  latitude,
  longitude = NULL,
  n = NULL,
  uz = 2,
  Rs = NULL,
  Rn = NULL,
  G = NULL,
  altitude = NULL,
  z = NULL,
  Krs = 0.16,
  Tmax = NULL,
  Tmin = NULL,
  albedo = 0.23,
  as = 0.25,
  bs = 0.5,
  ETins = NULL,
  ETr24 = NULL,
  wmo = NULL,
  airport = NULL,
  map = NULL,
  surface = "grass",
  period = "daytime"
)

# Default S3 method
ETohr(
  Tmean,
  RH,
  DOY,
  Lz,
  t1 = 1,
  time,
  Lm,
  latitude,
  longitude = NULL,
  n = NULL,
  uz = 2,
  Rs = NULL,
  Rn = NULL,
  G = NULL,
  altitude = NULL,
  z = NULL,
  Krs = 0.16,
  Tmax = NULL,
  Tmin = NULL,
  albedo = 0.23,
  as = 0.25,
  bs = 0.5,
  ETins = NULL,
  ETr24 = NULL,
  wmo = NULL,
  airport = NULL,
  map = NULL,
  surface = "grass",
  period = "daytime"
)

Arguments

Tmean

Numeric. Hourly Mean Air Temperature in degree Celsius

RH

Hourly Mean Relative Humidity in percent

DOY

Numeric or Date [YYYY-mm-dd]. Day of the Year. If you give data in the form of date [YYYY-mm-dd], it will be converted to DOY

Lz

longitude of the centre of the local time zone [degrees west of Greenwich]. MUST BE POSITIVE Lz = 0 degrees for Greenwich.

t1

The length of the calculation period in hour; 1 for hour, 0.5 for 30 minutes 0.25 for 15 minutes.

time

The midpoint of the time of measurement[hour]; for example time is 12.5 for a period between 12:00 and 13:00

Lm

longitude of the measurement site [degrees west of Greenwich]. MUST BE POSITIVE

latitude

geographical coordinates in decimal degrees. It should be negative for southern hemisphere

longitude

The longitude of the measurement site i.e. geographical coordinates in decimal degrees for the weather station. It should be negative for West and positive for East.

n

Sunshine hours

uz

The wind speed in m/s

Rs

solar radiation [MJ/m2/hr]

Rn

net radiation [MJ/m2/hr]

G

Daily Soil heat flux in W/m2

altitude

The elevation of the weather station, in metres

z

The height above surface where wind speed was measured, in metres.

Krs

Relationship between the fraction of extraterrestrial radiation that reaches the earth's surface, Rs/Ra, and the air temperature difference Tmax - Tmin for interior (Krs = 0.16) and coastal (Krs = 0.19) regions

Tmax

Numeric. Maximum air Temperature in degree Celsius

Tmin

Numeric. Minimum air Temperature in degree Celsius

albedo

A RasterLayer data that has albedo values. You can also provide file path or location on your computer

as

Regression constant, intercept, expressing the fraction of extra-terrestrial radiation reaching the earth on overcast days (n = 0); either a calibrated value or as = 0.25 can be used.

bs

regression constant, slope, expressing the fraction of extraterrestrial radiation reaching the earth on overcast days (n = 0); either a calibrated value or bs = 0.5 can be used.

ETins

model. It takes ET model from sebal or tseb and retrieve instantaneous ET estimates. This helps to estimate ET24, T24, E24, EF, Kc, EFs, EFc

ETr24

daily grass or alfalfa reference Evapotranspiration (mm/day)

wmo

numeric. World Meteorological Organization weather station code. You can use the Worldwide Station List at https://www.wunderground.com/ [Accessed on 2016-5-6] or https://www.wetterzentrale.de/ [Accessed on 2016-5-6]

airport

numeric. IATA or ICAO code. They are alphabetically listed at https://en.wikipedia.org/wiki/List_of_airports_by_IATA_code:_A [Accessed on 2016-5-6]

map

spatial raster data for inverse distance interpolation

surface

character. It is either "grass" or "alfalfa"

period

character. Time of day for the calculation: "daytime" or "nighttime". The default is "daytime".

Value

ETo:

hour reference ETo [mm/hr]

ETa:

hour actual ET. It is not NULL if EF has a numeric value

ETc:

hour crop ET. It is not NULL if Kc has a numeric value

Rn:

Net Radiation [W/m2]

Rnwm_2:

Net Radiation [MJ/m2/hr]

y:

psychrometric constant [kPa/degree Celsius]

slope:

slope of saturation vapour pressure curve [kPa/degree Celsius]

vpd:

Vapour pressure deficit[kPa]

Details

sebal, sebi,sebs,sseb, tseb

References

ALLEN, R. G., PEREIRA, L. S., RAES, D., & SMITH, M. 1998. Crop Evapotranspiration (guidelines for computing crop water requirements) FAO Irrigation and Drainage Paper No. 56: FAO.

Jackson, R. D., Hatfield, J. L., Reginato, R. J., Idso, S. B., & Pinter Jr, P. J. (1983). Estimation of daily evapotranspiration from one time-of-day measurements. Agricultural Water Management, 7(1-3), 351-362. doi: http://dx.doi.org/10.1016/0378-3774(83)90095-1

French, A. N., Fitzgerald, G., Hunsaker, D., Barnes, E., Clarke, T., Lesch, S., . . . Pinter, P. (2005). Estimating spatially distributed cotton water use from thermal infrared aerial imagery. Paper presented at the World Water Congress 2005: Impacts of Global Climate Change - Proceedings of the 2005 World Water and Environmental Resources Congress, Reston, Va.

Colaizzi, P. D., Kustas, W. P., Anderson, M. C., Agam, N., Tolk, J. A., Evett, S. R., . . . O'Shaughnessy, S. A. (2012). Two-source energy balance model estimates of evapotranspiration using component and composite surface temperatures. Advances in Water Resources, 50, 134-151. doi: http://dx.doi.org/10.1016/j.advwatres.2012.06.004

See also

Author

George Owusu

Examples

# \donttest{
#FAO56 Example 19 14:00-15:00 h using Cd = 0.24 instead 0.34 in FAO56
ET1hr=ETohr(Tmean=38,RH=52,DOY=274,Lz=15,t1=1,time=14.5,Lm=16.25,latitude=16.22,uz=3.3,Rs=2.450)
#> Computing ETr
ET1hr$ETo
#> [1] 0.655908
# }