A 8.5 RCP downscaled Weather data of Kumasi from 2016 to 2050

A dataset containing Rain, maximum and minimum temperature, wind speed, sunshine hours, and minimum and maximum Relative Humidity. The RCP 8.5 assumes that green house gas will rise throughout 21st Century.

Format

A data frame with 12784 rows and 8 variables:

DOY

Day of the Year in dates

Tmax

Maximum Temperature [oC]

Tmin

Minimum Temperature [oC]

RHmin

Minimum Relative Humidity [per cent]

RHmax

Maximum Relative Humidity [per cent]

n

sunshine hours

uz

wind speed [m/s]

P

Daily precipitation [mm]

Author

George Owusu